Renaming to pool2d
