+++ src/nn/pool2d_avg.rs
