+++ src/tensor_ops/mod.rs
@@ -196,6 +196,6 @@ mod conv;
 pub use conv::*;
 
 #[cfg(feature = "nightly")]
-mod pool;
+mod pool2d;
 #[cfg(feature = "nightly")]
-pub use pool::*;
+pub use pool2d::*;
